[dataprotect] BM-18308 fix: sds restore with SDS enabled: user password is not always user login, obviously..

+++ parent/dataprotect/net.bluemind.dataprotect.mailbox/src/net/bluemind/dataprotect/mailbox/MboxRestoreService.java
@@ -17,7 +17,7 @@
  * END LICENSE
  */
 
-package net.bluemind.dataprotect.mailbox.internal;
+package net.bluemind.dataprotect.mailbox;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -40,6 +40,7 @@ import org.slf4j.LoggerFactory;
 
 import com.google.common.collect.Lists;
 
+import net.bluemind.backend.cyrus.Sudo;
 import net.bluemind.backend.cyrus.partitions.CyrusPartition;
 import net.bluemind.backend.mail.api.IMailboxFolders;
 import net.bluemind.backend.mail.api.IMailboxFoldersByContainer;
@@ -50,8 +51,10 @@ import net.bluemind.common.cache.persistence.CacheBackingStore;
 import net.bluemind.config.InstallationId;
 import net.bluemind.config.Token;
 import net.bluemind.core.api.fault.ServerFault;
+import net.bluemind.core.container.api.IContainerManagement;
 import net.bluemind.core.container.model.ItemIdentifier;
 import net.bluemind.core.container.model.ItemValue;
+import net.bluemind.core.container.model.acl.Verb;
 import net.bluemind.core.context.SecurityContext;
 import net.bluemind.core.rest.BmContext;
 import net.bluemind.core.rest.IServiceProvider;
@@ -69,8 +72,13 @@ import net.bluemind.dataprotect.service.IDPContext;
 import net.bluemind.dataprotect.service.IDPContext.ITool;
 import net.bluemind.dataprotect.service.IDPContext.IToolConfig;
 import net.bluemind.dataprotect.service.IDPContext.IToolSession;
+import net.bluemind.directory.api.BaseDirEntry.Kind;
+import net.bluemind.directory.api.DirEntry;
 import net.bluemind.directory.api.IDirEntryMaintenance;
+import net.bluemind.directory.api.IDirectory;
 import net.bluemind.domain.api.Domain;
+import net.bluemind.group.api.IGroup;
+import net.bluemind.group.api.Member;
 import net.bluemind.imap.Flag;
 import net.bluemind.imap.FlagsList;
 import net.bluemind.imap.IMAPException;
@@ -97,10 +105,10 @@ public class MboxRestoreService {
 		Replace, Subfolder
 	}
 
-	private static final Set<String> defaultFolders = new HashSet<>(DefaultFolder.USER_FOLDERS_NAME);
+	private static final Set<String> defaultUserFolders = new HashSet<>(DefaultFolder.USER_FOLDERS_NAME);
 
 	public MboxRestoreService() {
-		defaultFolders.add("INBOX");
+		defaultUserFolders.add("INBOX");
 	}
 
 	public static final SecurityContext as(String uid, String domainContainerUid) throws ServerFault {
@@ -146,7 +154,12 @@ public class MboxRestoreService {
 
 		SystemConf sysconf = LocalSysconfCache.get();
 		if (sysconf.isArchiveKindSds()) {
-			restoreSds(domain, dpg, mbox, boxFsFolders, mode, monitor);
+			try {
+				restoreSds(domain, dpg, mbox, boxFsFolders, mode, monitor);
+			} catch (ServerFault sf) {
+				monitor.end(false, "finished", "{ \"status\": \"Server Error:" + sf.getMessage() + "\" }");
+				throw sf;
+			}
 		} else {
 			restoreRsync(domain, dpg, mailPart, mbox, boxFsFolders, mode, monitor);
 		}
@@ -178,7 +191,7 @@ public class MboxRestoreService {
 
 			ITask taskApi = sp.instance(ITask.class, tr.id);
 			TaskUtils.wait(sp, tr);
-			taskApi.getCurrentLogs().stream().forEach(l -> monitor.log(l));
+			taskApi.getCurrentLogs().stream().forEach(monitor::log);
 			break;
 		case Subfolder:
 			if (mbox.value.type == Type.mailshare) {
@@ -230,12 +243,72 @@ public class MboxRestoreService {
 		logger.info("ending task with mon {}", monitor);
 	}
 
+	/**
+	 * Returns one user with write right on the mailshare
+	 * 
+	 * @param domainUid
+	 * @param mbox
+	 * @return
+	 */
+	private String getMailshareWriter(String domainUid, ItemValue<Mailbox> mbox) {
+		IServiceProvider sp = ServerSideServiceProvider.getProvider(SecurityContext.SYSTEM);
+		IContainerManagement mailboxAclsService = sp.instance(IContainerManagement.class, "mailbox:acls-" + mbox.uid);
+		List<String> writers = mailboxAclsService.getAccessControlList().stream()
+				.filter(entity -> entity.verb.can(Verb.Write)).map(entry -> entry.subject).collect(Collectors.toList());
+		if (writers.isEmpty()) {
+			return null;
+		}
+		IDirectory dir = sp.instance(IDirectory.class, domainUid);
+		List<ItemValue<DirEntry>> entries = dir.getMultiple(writers);
+		Optional<ItemValue<DirEntry>> dirEntry = entries.stream().filter(de -> de.value.kind == Kind.USER).findFirst();
+		String entryUid = null;
+		if (!dirEntry.isPresent()) {
+			IGroup groupService = sp.instance(IGroup.class, domainUid);
+			List<ItemValue<DirEntry>> groups = entries.stream().filter(de -> de.value.kind == Kind.GROUP)
+					.collect(Collectors.toList());
+			for (int i = 0; i < groups.size(); i++) {
+				ItemValue<DirEntry> g = groups.get(i);
+				List<Member> members = groupService.getExpandedUserMembers(g.uid);
+				if (!members.isEmpty()) {
+					entryUid = members.get(0).uid;
+					break;
+				}
+			}
+		} else {
+			entryUid = dirEntry.get().uid;
+		}
+		return entryUid;
+	}
+
 	private void restoreSds(ItemValue<Domain> domain, DataProtectGeneration dpg, ItemValue<Mailbox> mbox,
 			BoxFsFolders boxFsFolders, Mode mode, IServerTaskMonitor monitor) {
-		SecurityContext backUserContext = as(mbox.uid, domain.uid);
-		BmContext live = ServerSideServiceProvider.getProvider(backUserContext).getContext();
-		String subtree = IMailReplicaUids.subtreeUid(domain.uid, mbox);
-		IMailboxFolders foldersService = live.getServiceProvider().instance(IMailboxFoldersByContainer.class, subtree);
+		BmContext live;
+		Set<String> defaultFolders;
+		boolean isMailshare = mbox.value.type == Type.mailshare;
+		String inboxName = isMailshare ? mbox.value.name : "INBOX";
+
+		if (isMailshare) {
+			defaultFolders = new HashSet<>(DefaultFolder.MAILSHARE_FOLDERS_NAME.stream()
+					.map(f -> mbox.value.name + "/" + f).collect(Collectors.toList()));
+			defaultFolders.add(mbox.value.name);
+			// Find another used with write permissions
+			String entryUid = getMailshareWriter(domain.uid, mbox);
+			logger.info("Found entry with write permissions on {}: {}", mbox, entryUid);
+			if (entryUid == null || entryUid.isEmpty()) {
+				monitor.end(false, "finished",
+						"{ \"status\": \"Unable to find a user with write permissions on mailshare " + mbox.value.name
+								+ "\" }");
+				throw new ServerFault("Unable to find a user with write permissions on mailshare " + mbox.value.name);
+			}
+			live = ServerSideServiceProvider.getProvider(SecurityContext.SYSTEM).getContext()
+					.su("sds-restore-" + UUID.randomUUID().toString(), entryUid, domain.uid);
+		} else {
+			defaultFolders = new HashSet<>(defaultUserFolders);
+			live = ServerSideServiceProvider.getProvider(as(mbox.uid, domain.uid)).getContext();
+		}
+
+		IMailboxFolders foldersService = live.getServiceProvider().instance(IMailboxFoldersByContainer.class,
+				IMailReplicaUids.subtreeUid(domain.uid, mbox));
 		IServer srvApi = ServerSideServiceProvider.getProvider(SecurityContext.SYSTEM).instance(IServer.class,
 				InstallationId.getIdentifier());
 
@@ -273,15 +346,27 @@ public class MboxRestoreService {
 
 		// clean / remove folders before injecting backup
 		if (mode == Mode.Replace) {
-			for (ItemValue<MailboxFolder> liveFolder : foldersService.all()) {
+			// We need to remove child firsts
+			List<ItemValue<MailboxFolder>> liveFolders = foldersService.all(); //
+			liveFolders.sort((a, b) -> Long.valueOf(a.value.fullName.chars().filter(ch -> ch == '/').count())
+					.compareTo(b.value.fullName.chars().filter(ch -> ch == '/').count()));
+			Collections.reverse(liveFolders);
+			for (ItemValue<MailboxFolder> liveFolder : liveFolders) {
 				if (!defaultFolders.contains(liveFolder.value.fullName)) {
 					try {
+						logger.info("Removing {}", liveFolder);
 						foldersService.deleteById(liveFolder.internalId);
 					} catch (ServerFault sf) {
 						logger.error("Unable to remove folder {}: {}", liveFolder.value.fullName, sf.getMessage());
 					}
 				} else {
 					try {
+						if (isMailshare && !liveFolder.value.fullName.startsWith(mbox.value.name)) {
+							// Just a "useless" safety, we don't want to clear the write user mailbox by
+							// error.
+							continue;
+						}
+						logger.info("Removing messages of {}", liveFolder);
 						foldersService.removeMessages(liveFolder.internalId);
 					} catch (ServerFault sf) {
 						logger.error("Unable to remove messages from folder {}: {}", liveFolder.value.fullName,
@@ -294,16 +379,25 @@ public class MboxRestoreService {
 		// Now, restore the folders using bluemind api
 		// We went to sort "INBOX" first
 		List<CyrusSdsBackupFolder> folders = sdsBackupMailbox.getFolders();
-		folders.sort((a, b) -> a.fullNameWithoutInbox().compareTo(b.fullNameWithoutInbox()));
+
+		// sort by hierarchy depth
+		folders.sort((a, b) -> Long.valueOf(a.fullNameWithoutInbox(inboxName).chars().filter(ch -> ch == '/').count())
+				.compareTo(b.fullNameWithoutInbox(inboxName).chars().filter(ch -> ch == '/').count()));
 
 		for (CyrusSdsBackupFolder folder : folders) {
 			MailboxFolder createdfolder = new MailboxFolder();
 			// restoredFolderName is "restored-blibla" and folder is the "source folder"
 			// (from the backup)
 			if (mode == Mode.Subfolder) {
-				createdfolder.fullName = boxFsFolders.restoreFolderName
-						+ (!folder.fullNameWithoutInbox().isEmpty() ? ("/" + folder.fullNameWithoutInbox()) : "");
-				createdfolder.name = folder.name().equals("INBOX") ? boxFsFolders.restoreFolderName : folder.name();
+				createdfolder.name = folder.name().equals(inboxName) ? boxFsFolders.restoreFolderName : folder.name();
+				String fullName = boxFsFolders.restoreFolderName + (!folder.fullNameWithoutInbox(inboxName).isEmpty()
+						? ("/" + folder.fullNameWithoutInbox(inboxName))
+						: "").replace("//", "/");
+				if (isMailshare) {
+					createdfolder.fullName = mbox.value.name + "/" + fullName;
+				} else {
+					createdfolder.fullName = fullName;
+				}
 			} else if (mode == Mode.Replace) {
 				createdfolder.fullName = folder.fullName();
 				createdfolder.name = folder.name();
@@ -312,35 +406,54 @@ public class MboxRestoreService {
 				monitor.end(false, "finished", "{ \"status\": \"Unsupported restore mode\" }");
 				return;
 			}
-			logger.debug("will create folder: fullName: {} name: {}", createdfolder.fullName, createdfolder.name);
-			try {
-				ItemIdentifier createAck = foldersService.createBasic(createdfolder);
-				logger.debug("live folder {} createack: {}", createdfolder, createAck);
-			} catch (ServerFault sf) {
-				logger.error("folder {} creation failed", sf.getMessage());
+			if (!defaultFolders.contains(createdfolder.fullName)) {
+				logger.info("Create folder: {}", createdfolder);
+				monitor.log("Create folder: {}", createdfolder);
+				try {
+					ItemIdentifier createAck = foldersService.createBasic(createdfolder);
+					logger.debug("live folder {} createack: {}", createdfolder, createAck);
+				} catch (ServerFault sf) {
+					logger.error("unable to create folder '{}': {}", createdfolder, sf.getMessage());
+					monitor.log("unable to create folder '{}': {}", createdfolder, sf.getMessage());
+					continue;
+				}
+			} else {
+				logger.debug("{} is a defaultFolder, so it will not be created", createdfolder);
 			}
 
-			try (StoreClient sc = new StoreClient(source.value.address(), 1143, mbox.value.name + "@" + domain.uid,
-					mbox.value.name)) {
-				if (!sc.login()) {
-					throw new ServerFault("error logging in");
+			if (mbox.value.type == Type.mailshare) {
+				try (StoreClient sc = new StoreClient(source.value.address(), 1143, "admin0", Token.admin0())) {
+					if (!sc.login()) {
+						throw new ServerFault("Error logging in as admin0");
+					}
+					restoreSdsMessages(live, sc, folder, createdfolder.fullName + "@" + domain.uid, monitor);
 				}
-				restoreSdsMessages(live, sc, folder, createdfolder, monitor);
+			} else {
+				try (Sudo sudo = Sudo.forLogin(mbox.value.name, domain.uid)) {
+					try (StoreClient sc = new StoreClient(source.value.address(), 1143,
+							mbox.value.name + "@" + domain.uid, sudo.context.getSessionId())) {
+						if (!sc.login()) {
+							throw new ServerFault("error logging in to " + source.value.address() + ":1143 as "
+									+ mbox.value.name + "@" + domain.uid);
+						}
+						restoreSdsMessages(live, sc, folder, createdfolder.fullName, monitor);
+					}
+				}
 			}
 		}
 		monitor.end(true, "finished", "{ \"status\": \"not_implemented\" }");
 		logger.info("ending task with mon {}", monitor);
 	}
 
-	private void restoreSdsMessages(BmContext context, StoreClient sc, CyrusSdsBackupFolder folder,
-			MailboxFolder mboxFolder, IServerTaskMonitor monitor) {
+	private void restoreSdsMessages(BmContext context, StoreClient sc, CyrusSdsBackupFolder folder, String mboxFolder,
+			IServerTaskMonitor monitor) {
 		MessageBodyObjectStore sds = new MessageBodyObjectStore(context);
 
 		FlagsList flags = new FlagsList();
 		flags.add(Flag.BMARCHIVED);
 		flags.add(Flag.SEEN);
 		long allMessagesCount = folder.messageCount();
-		monitor.log(String.format("Restoring %d messages of folder %s", allMessagesCount, folder.fullName()));
+		monitor.log("Restoring {} messages of folder {}", allMessagesCount, folder.fullName());
 		AtomicLong restorationCounter = new AtomicLong(0);
 		List<List<CyrusSdsBackupMessage>> partitioned = Lists.partition(folder.messages(), 32);
 		partitioned.stream().forEach(msgList -> {
@@ -351,13 +464,14 @@ public class MboxRestoreService {
 				// msgList and paths must be in the same order, this is critical
 				CyrusSdsBackupMessage msg = msgList.get(idx++);
 				try (InputStream instream = Files.newInputStream(p)) {
-					int added = sc.append(mboxFolder.fullName, instream, flags, msg.date);
+					int added = sc.append(mboxFolder, instream, flags, msg.date);
 					if (added <= 0) {
 						logger.error("Unable to inject message {}", p);
-					}
-					long currentCount = restorationCounter.incrementAndGet();
-					if (currentCount == allMessagesCount || (currentCount % 1000) == 0) {
-						monitor.log(String.format("Restored %d/%d messages", currentCount, allMessagesCount));
+					} else {
+						long currentCount = restorationCounter.incrementAndGet();
+						if (currentCount == allMessagesCount || (currentCount % 1000) == 0) {
+							monitor.log("Restored {}/{} messages", currentCount, allMessagesCount);
+						}
 					}
 				} catch (IOException e) {
 					logger.error("unable to open downloaded message {}", p, e);
@@ -369,6 +483,7 @@ public class MboxRestoreService {
 					} catch (IOException e) {
 					}
 				}
+				monitor.log("Restored {}/{} messages", restorationCounter.get(), allMessagesCount);
 			}
 		});
 
@@ -409,9 +524,7 @@ public class MboxRestoreService {
 		NCUtils.exec(nc, String.format("mkdir -p '%s'", rootPath));
 		NCUtils.exec(nc, String.format("chown cyrus:mail '%s'", rootPath));
 		NCUtils.exec(nc, String.format("chmod 700 '%s'", rootPath));
-		path.forEach(f -> {
-			session.restoreOneFolder(mailPartId, f, rootPath);
-		});
+		path.forEach(f -> session.restoreOneFolder(mailPartId, f, rootPath));
 	}
 
 }
